+++ commands/dbaas/mongo/cluster/list.go
@@ -3,8 +3,6 @@ package cluster
 import (
 	"context"
 
-	"github.com/spf13/viper"
-
 	"github.com/ionos-cloud/ionosctl/v6/pkg/constants"
 	"github.com/ionos-cloud/ionosctl/v6/pkg/core"
 )
@@ -21,18 +19,7 @@ func ClusterListCmd() *core.Command {
 		PreCmdRun: core.NoPreRun,
 		CmdRun: func(c *core.CommandConfig) error {
 			c.Printer.Verbose("Getting Clusters...")
-			name := viper.GetString(core.GetFlagName(c.NS, constants.FlagName))
-			var limitPtr *int32 = nil
-			if f := core.GetFlagName(c.NS, constants.FlagMaxResults); viper.IsSet(f) {
-				limit := viper.GetInt32(f)
-				limitPtr = &limit
-			}
-			var offsetPtr *int32 = nil
-			if f := core.GetFlagName(c.NS, constants.FlagOffset); viper.IsSet(f) {
-				offset := viper.GetInt32(f)
-				offsetPtr = &offset
-			}
-			clusters, _, err := c.DbaasMongoServices.Clusters().List(name, limitPtr, offsetPtr)
+			clusters, err := Clusters(FilterPaginationFlags(c), FilterNameFlags(c))
 			if err != nil {
 				return err
 			}